lara: adopt burn flag in TR1
This updates TR1 to use the same approach as TR2 in tracking
Lara's burn status, saving it separately in savegame files to work
independently of having effect saving on. In addition, this merges
Lara's logic for catching fire and putting flames out into TRX.

Resolves #3157.
