--- src/tr1/game/lara/misc.c
@@ -1,28 +1,0 @@
-#include "game/effects.h"
-#include "global/vars.h"
-
-void Lara_CatchFire(void)
-{
-    const int16_t effect_num = Effect_Create(g_LaraItem->room_num);
-    if (effect_num != NO_EFFECT) {
-        EFFECT *const effect = Effect_Get(effect_num);
-        effect->frame_num = 0;
-        effect->object_id = O_FLAME;
-        effect->counter = -1;
-    }
-}
-
-void Lara_Extinguish(void)
-{
-    // put out flame objects
-    int16_t effect_num = Effect_GetActiveNum();
-    while (effect_num != NO_EFFECT) {
-        EFFECT *const effect = Effect_Get(effect_num);
-        const int16_t next_effect_num = effect->next_active;
-        if (effect->object_id == O_FLAME && effect->counter < 0) {
-            effect->counter = 0;
-            Effect_Kill(effect_num);
-        }
-        effect_num = next_effect_num;
-    }
-}
