+++ src/tr1/game/savegame/savegame_bson.c
@@ -814,6 +814,7 @@ static bool M_LoadLara(
         JSON_ObjectGetInt(lara_obj, "death_count", lara->death_timer);
     lara->current_active =
         JSON_ObjectGetInt(lara_obj, "current_active", lara->current_active);
+    lara->burn = JSON_ObjectGetBool(lara_obj, "burn", lara->burn);
 
     lara->hit_effect_count =
         JSON_ObjectGetInt(lara_obj, "hit_effect_count", lara->hit_effect_count);
@@ -1296,6 +1297,7 @@ static JSON_OBJECT *M_DumpLara(LARA_INFO *lara)
     JSON_ObjectAppendInt(lara_obj, "dive_count", lara->dive_timer);
     JSON_ObjectAppendInt(lara_obj, "death_count", lara->death_timer);
     JSON_ObjectAppendInt(lara_obj, "current_active", lara->current_active);
+    JSON_ObjectAppendBool(lara_obj, "burn", lara->burn);
 
     JSON_ObjectAppendInt(lara_obj, "hit_effect_count", lara->hit_effect_count);
     JSON_ObjectAppendInt(
