+++ src/tr1/meson.build
@@ -144,7 +144,6 @@ sources = [
   'game/lara/control.c',
   'game/lara/draw.c',
   'game/lara/look.c',
-  'game/lara/misc.c',
   'game/lara/state.c',
   'game/level.c',
   'game/los.c',
