+++ src/tr2/game/lara/misc.c
@@ -416,42 +416,3 @@ void Lara_WaterCurrent(COLL_INFO *const coll)
 
     coll->old = item->pos;
 }
-
-void Lara_CatchFire(void)
-{
-    if (g_Lara.burn || g_Lara.water_status == LWS_CHEAT) {
-        return;
-    }
-
-    const int16_t effect_num = Effect_Create(g_LaraItem->room_num);
-    if (effect_num == NO_EFFECT) {
-        return;
-    }
-
-    EFFECT *const effect = Effect_Get(effect_num);
-    effect->frame_num = 0;
-    effect->object_id = O_FLAME;
-    effect->counter = -1;
-    g_Lara.burn = true;
-}
-
-void Lara_Extinguish(void)
-{
-    if (!g_Lara.burn) {
-        return;
-    }
-
-    g_Lara.burn = false;
-
-    // put out flame objects
-    int16_t effect_num = Effect_GetActiveNum();
-    while (effect_num != NO_EFFECT) {
-        EFFECT *const effect = Effect_Get(effect_num);
-        const int16_t next_effect_num = effect->next_active;
-        if (effect->object_id == O_FLAME && effect->counter < 0) {
-            effect->counter = 0;
-            Effect_Kill(effect_num);
-        }
-        effect_num = next_effect_num;
-    }
-}
